yatube/posts/forms.py
from django import forms
from django.utils.translation import gettext_lazy as _

from posts.models import Comment, Post


class PostForm(forms.ModelForm):
    class Meta:
        model = Post
        fields = ("text", "group", "image")
        labels = {
            "text": "Текст сообщения",
            "group": "Сообщество",
            "image": "Картинки поста",
        }


class CommentForm(forms.ModelForm):
    class Meta:
        model = Comment
        fields = ("text",)
        labels = {"text": "Текст комментария"}
